+++ CHANGELOG.md
@@ -6,6 +6,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Changelog plugin automatically applied to `project`
+- Feature matrix and Readme generation task for a `project` plugin
+
+### Changed
+- Remove node plugin and make node binaries available by default
+- Use default webpack distribution path
+- `ru.mipt.npm.base` -> `ru.mipt.npm.project`
+- Move publishing out of general extension and apply it to project plugin instead.
+
+### Deprecated
+
+### Removed
+- Node plugin
+
+### Fixed
+
+### Security
+## [0.6.0]
 
 ### Added
 - Migrate to kotlin 1.4.0